backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
channels>=4.1,<5.0
daphne>=4.1,<5.0
django-cors-headers>=4.3,<5.0
opencv-python>=4.9,<5.0
mediapipe>=0.10,<0.11
numpy>=1.24,<3.0
python-socketio>=5.11,<6.0
